manage: reindexing tasks
* Adds new manager `inspirehep` for running INSPIRE
  site specific commands.

* Adds two new commands `inveniomanage inspirehep reindex` and
  `inveniomanage inspirehep reindexhp` for re-indexing records
  and Holding Pen records respectively.

* Moves `create_indicies` command to `inspirehep` manager.

* Updates `inveniomanage migrator populate` command options, by
  simplifying the interface and removing broken code.
