+++ inspirehep/__init__.py
@@ -20,4 +20,6 @@
 # granted to it by virtue of its status as an Intergovernmental Organization
 # or submit itself to any jurisdiction.
 
+from .receivers import *
+
 __import__('pkg_resources').declare_namespace(__name__)
