+++ inspirehep/modules/migrator/__init__.py
@@ -16,5 +16,3 @@
 # You should have received a copy of the GNU General Public License
 # along with INSPIRE; if not, write to the Free Software Foundation, Inc.,
 # 59 Temple Place, Suite 330, Boston, MA 02111-1307, USA.
-
-from .receivers import *
